requirements/test.txt
-r production.txt

django-extensions==3.1.5
django-grappelli==2.15.3
django-guardian==2.4.0
mock==4.0.3
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
